app/requirements.txt
asgiref==3.2.7
atomicwrites==1.4.0
attrs==19.3.0
certifi==2020.4.5.1
chardet==3.0.4
colorama==0.4.3
defusedxml==0.6.0
Django==3.0.7
django-allauth==0.42.0
django-rest-auth==0.9.5
djangorestframework==3.11.0
idna==2.9
importlib-metadata==1.6.1
more-itertools==8.3.0
oauthlib==3.1.0
packaging==20.4
pip==20.1.1
pluggy==0.13.1
py==1.8.1
pyparsing==2.4.7
python3-openid==3.1.0
pytz==2020.1
requests==2.23.0
requests-oauthlib==1.3.0
setuptools==47.1.1
six==1.15.0
sqlparse==0.3.1
urllib3==1.25.9
wcwidth==0.2.3
zipp==3.1.0
